Break up shim.js
