+++ src/initializers/screen-dismissal.js
@@ -1,0 +1,3 @@
+export const dismissingScreenListener = { current: undefined };
+
+global.__rainbowDismissScreen = () => dismissingScreenListener.current?.();
