+++ src/screens/SendSheetEOA.js
@@ -10,13 +10,13 @@ import React, { useCallback, useEffect, useRef, useState } from 'react';
 import { InteractionManager, Keyboard } from 'react-native';
 import { isIphoneX } from 'react-native-iphone-x-helper';
 import { useDispatch } from 'react-redux';
-import { dismissingScreenListener } from '../../shim';
 import SendSheet, {
   useSendAddressValidation,
   useShowAssetFlags,
 } from '../components/send/SendSheet';
 import { createSignableTransaction, estimateGasLimit } from '../handlers/web3';
 import AssetTypes from '../helpers/assetTypes';
+import { dismissingScreenListener } from '../initializers/screen-dismissal';
 import { sendTransaction } from '../model/wallet';
 import { useNavigation } from '../navigation/Navigation';
 import { SEND_TRANSACTION_ERROR_MESSAGE } from '@cardstack/constants';
